backend/requirements.txt
# Runtime deps for backend (subset of root pyproject.toml for Docker build context)
fastapi>=0.128.6
jinja2>=3.1.6
uvicorn>=0.40.0
pydantic-settings>=2.12.0
python-dotenv>=1.2.1
google-genai>=1.0.0
sqlmodel>=0.0.32
# PostgreSQL driver for SQLAlchemy when using postgresql:// in Docker Compose
psycopg2-binary>=2.9.0
# For running tests inside the container (incl. Tectonic integration tests)
pytest>=9.0.2
pytest-asyncio>=1.3.0
